Provide context with errors to help users
We currently flatten the serde_json errors into strings quite
early on, loosing the ability to do structured reporting on the
underlying problem; this particularly shows up during API version
upgrades or when working on CRDs. Pulling out some context from the
bytes we're parsing helps (in the JSON case at least) without requiring
the operation to be repeatable with debug! logging.