Preserve the failed input in decode errors
This does the other requested extension, preserving the
full input in decode errors by layering a custom failure on top.

I've laid the ground work for migrating other errors into the client
crate, but not done so to keep the delta reviewable.

Of course, it wasn't as clean as desired, but I chose not to fully
implement Fail by hand, instead linking to the requested enhancement
upstream in failure.